feat: mapper

+++ src/main/java/br/kauesoares/dto/Input.java
@@ -20,6 +20,6 @@ public final class Input {
     public float kmHome;
 
     public boolean hasLast;
-    public float minutesSince;
+    public double minutesSince;
     public float kmLast;
 }
